Return usize::MAX for recursive structures
This is technically correct, since a structure like
`struct A { a: Box<A> }` will never be instanciated by arbitrary.
A structure such as `struct A { a: Option<Box<A>> }` will still
return a size_hint of (0,None) because of the `or` in the implementation
of Option.

This also switches the `and` util to use saturating arithmetric to avoid
overflows linked to reaching usize::MAX